Support custom baseurl and jekyll serve. Update Gems.
This updated to use newer version of Gems-dependencies.
There are still warning and deprecated shouting out when doing "bundle
exec jekyll serve" but it works at the end for this commit for now.

To get the path of asset instead of using normal way of `{% javascript
app.js %}`, we use `<script type="text/javascript" src="{{ site.baseurl
}}{% asset_path app.js %}"></script>` instead as we need to prepend
site.baseurl.

baseurl and url are added to _config.yml.

So in short, when in development just hit `bundle exec jekyll serve`. It
will provide live compile, and you can see immediate update on
http://localhost:4000.